requirements.txt
argon2-cffi==16.2.0
cffi==1.8.3
pkg-resources==0.0.0
pycparser==2.14
PyNaCl==1.0.1
six==1.10.0
